refactor: 회원가입 페이지 리팩토링
- MetaData 추가 및 SSR 적용
